Adjust zsh config for headless Raspbian
Removing references to vscode and GUI programs such as gitx or gitk
and removing omz plugins so it can load faster. In the future,
I will remove oh-my-zsh as a whole because it takes a lot of toll
on a small system such as a Raspberry Pi. When I changed the shell
to bash it improved the speed significantly.

Also force color prompt in bashrc