profile(cachyos): Enable Vietnamese input using fcitx5-bamboo on Wayland
This comes with 2 caveats:

1. Non-Chromium apps that are built on Electron:

Caveat: Any apps that are built in electron that needs to be enabled
to type in Vietnamese must be installed with the version installed with
System Electron instead of the provided binary for these configs to get
picked up.

e.g.: For Discord on AUR there are 2 official clients:
  - `discord` which uses the provided electron bin
  - `discord_arch_bin` which uses system provided electron

If I want to use Discord with Vietnamese input, I need to pick the 2nd
option.

2. For Chromium/Chrome browsers:

As documented in Arch Linux docs
(https://wiki.archlinux.org/title/Chromium#Making_flags_persistent),
these flags should be applied accordingly for what browsers you uses
(i.e. for Google Chrome use `chrome-`, for Chromium use
`chromium-`...)